pyproject.toml
[tool.poetry]
name = "wallet-pipeline"
version = "0.1.0"
description = ""
authors = ["tomfutago <tomfutago@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
duckdb = "^1.1.3"
fsspec = "^2024.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
